+++ buckifier/buckify_rocksdb.py
@@ -108,16 +108,6 @@ def get_tests(repo_path):
 
     return tests
 
-
-# Get gtest dir from Makefile
-def get_gtest_dir(repo_path):
-    for line in open(repo_path + "/Makefile"):
-        if line.strip().startswith("GTEST_DIR ="):
-            return line.split("=")[1].strip()
-    # if not found
-    exit_with_error("Unable to find GTEST_DIR in Makefile")
-
-
 # Parse extra dependencies passed by user from command line
 def get_dependencies():
     deps_map = {
@@ -151,13 +141,11 @@ def generate_targets(repo_path, deps_map):
     cc_files = get_cc_files(repo_path)
     # get tests from Makefile
     tests = get_tests(repo_path)
-    # get gtest dir
-    gtest_dir = get_gtest_dir(repo_path) + "/"
 
     if src_mk is None or cc_files is None or tests is None:
         return False
 
-    TARGETS = TARGETSBuilder("%s/TARGETS" % repo_path, gtest_dir)
+    TARGETS = TARGETSBuilder("%s/TARGETS" % repo_path)
 
     # rocksdb_lib
     TARGETS.add_library(
@@ -171,7 +159,10 @@ def generate_targets(repo_path, deps_map):
         src_mk.get("TEST_LIB_SOURCES", []) +
         src_mk.get("EXP_LIB_SOURCES", []) +
         src_mk.get("ANALYZER_LIB_SOURCES", []),
-        [":rocksdb_lib", ":rocksdb_third_party_gtest"])
+        [":rocksdb_lib"],
+        extra_external_deps=""" + [
+        ("googletest", None, "gtest"),
+    ]""")
     # rocksdb_tools_lib
     TARGETS.add_library(
         "rocksdb_tools_lib",
@@ -185,12 +176,6 @@ def generate_targets(repo_path, deps_map):
         src_mk.get("ANALYZER_LIB_SOURCES", [])
         + src_mk.get('STRESS_LIB_SOURCES', [])
         + ["test_util/testutil.cc"])
-    # rocksdb_third_party_gtest
-    TARGETS.add_library(
-        "rocksdb_third_party_gtest",
-        [gtest_dir + "gtest/gtest-all.cc"],
-        [],
-        [gtest_dir + "gtest/gtest.h"])
 
     print("Extra dependencies:\n{0}".format(json.dumps(deps_map)))
     # test for every test we found in the Makefile